jest-jasmine1 && jest-jasmine2
