+++ package.json
@@ -9,6 +9,8 @@
     "diff": "^2.1.1",
     "graceful-fs": "^4.1.3",
     "istanbul": "^0.4.2",
+    "jest-jasmine1": "^0.0.0",
+    "jest-jasmine2": "^0.0.0",
     "jest-mock": "^1.0.0",
     "jest-util": "^1.0.0",
     "jsdom": "^7.2.0",
