+++ packages/jest-jasmine1/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "jest-jasmine1",
+  "version": "0.0.0",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/facebook/jest.git"
+  },
+  "license": "BSD-3-Clause",
+  "main": "src/index.js",
+  "dependencies": {},
+  "scripts": {
+    "test": "jest"
+  },
+  "jest": {
+    "testEnvironment": "<rootDir>/../../src/environments/NodeEnvironment"
+  }
+}
