+++ packages/jest-jasmine1/src/__tests__/reporter-test.js
@@ -11,79 +11,69 @@
 
 jest.disableAutomock();
 
-describe('Jasmine2Reporter', function() {
+describe('JasmineReporter', function() {
+  // modules
   var JasmineReporter;
   var chalk;
+
+  // other variables
   var reporter;
 
   beforeEach(function() {
-    JasmineReporter = require('../Jasmine2Reporter');
+    JasmineReporter = require('../reporter');
     chalk = require('chalk');
 
     reporter = new JasmineReporter();
   });
 
-  describe('suites', function() {
-
-    pit('reports nested suites', function() {
-      var makeSpec = function(name) {
-        return {
-          fullName: name,
-          description: 'description',
-          failedExpectations: [],
-        };
-      };
-      reporter.suiteStarted({description: 'parent'});
-      reporter.suiteStarted({description: 'child'});
-      reporter.specDone(makeSpec('spec 1'));
-      reporter.suiteDone();
-      reporter.suiteStarted({description: 'child 2'});
-      reporter.specDone(makeSpec('spec 2'));
-      reporter.jasmineDone();
-
-      return reporter.getResults().then(function(runResults) {
-
-        var firstResult = runResults.testResults[0];
-        expect(firstResult.ancestorTitles[0]).toBe('parent');
-        expect(firstResult.ancestorTitles[1]).toBe('child');
-        var secondResult = runResults.testResults[1];
-        expect(secondResult.ancestorTitles[1]).toBe('child 2');
-
-      });
-    });
-
-  });
-
   describe('colorization', function() {
-
-    function getFailedResult(actualResult, expectedResult) {
-      var desc = 'comparing ' + actualResult + ' to ' + expectedResult;
+    function getRunner(item) {
       return {
-        fullName: desc,
-        description: desc,
-        failedExpectations: [
-          {
-            actual: actualResult,
-            expected: expectedResult,
-            matcherName: 'toBe',
-            passed: false,
-          },
-        ],
+        suites: function() {
+          return [
+            {
+              parentSuite: null,
+              specs: function() {
+                return [
+                  {
+                    results: function() {
+                      return {
+                        getItems: function() {
+                          return [item];
+                        },
+                      };
+                    },
+                  },
+                ];
+              },
+              suites: function() { return []; },
+            },
+          ];
+        },
       };
     }
 
-    function getExceptionResult(stack) {
-      return {
-        fullName: '',
-        description: '',
-        failedExpectations: [
-          {
-            matcherName: '',
-            stack: stack,
-            passed: false,
-          },
-        ],
-      };
+    function getExpectedRunner(actualResult, expectedResult, passed) {
+      return getRunner({
+        actual: actualResult,
+        expected: expectedResult,
+        message: '',
+        isNot: false,
+        matcherName: 'toBe',
+        passed: function() { return passed; },
+        trace: {},
+        type: 'expect',
+      });
+    }
+
+    function getExceptionRunner(message, passed) {
+      return getRunner({
+        passed: function() { return passed; },
+        trace: {
+          stack: message,
+        },
+        type: 'expect',
+      });
     }
 
     function errorize(str) {
@@ -95,9 +85,8 @@ describe('Jasmine2Reporter', function() {
     }
 
     pit('colorizes single-line failures using a per-char diff', function() {
-      var result = getFailedResult('foo', 'foobar');
-      reporter.specDone(result);
-      reporter.jasmineDone();
+      var runner = getExpectedRunner('foo', 'foobar', false);
+      reporter.reportRunnerResults(runner);
 
       return reporter.getResults().then(function(result) {
         var message = result.testResults[0].failureMessages[0];
@@ -109,9 +98,8 @@ describe('Jasmine2Reporter', function() {
     });
 
     pit('colorizes multi-line failures using a per-line diff', function() {
-      var result = getFailedResult('foo\nbar\nbaz', 'foo\nxxx\nbaz');
-      reporter.specDone(result);
-      reporter.jasmineDone();
+      var runner = getExpectedRunner('foo\nbar\nbaz', 'foo\nxxx\nbaz', false);
+      reporter.reportRunnerResults(runner);
 
       return reporter.getResults().then(function(result) {
         var message = result.testResults[0].failureMessages[0];
@@ -123,15 +111,15 @@ describe('Jasmine2Reporter', function() {
     });
 
     pit('colorizes exception messages', function() {
-      var result = getExceptionResult(
+      var runner = getExceptionRunner(
         'Error: foobar = {\n' +
         '      attention: "bar"\n' +
         '    }\n' +
         '    at Error (<anonymous>)\n' +
-        '    at Baz.js (<anonymous>)'
+        '    at Baz.js (<anonymous>)',
+        false
       );
-      reporter.specDone(result);
-      reporter.jasmineDone();
+      reporter.reportRunnerResults(runner);
 
       return reporter.getResults().then(function(result) {
         var message = result.testResults[0].failureMessages[0];
@@ -144,8 +132,6 @@ describe('Jasmine2Reporter', function() {
           '    at Baz.js (<anonymous>)'
         );
       });
-
     });
-
   });
 });
