+++ packages/jest-jasmine2/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "jest-jasmine2",
+  "version": "0.0.0",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/facebook/jest.git"
+  },
+  "license": "BSD-3-Clause",
+  "main": "src/index.js",
+  "dependencies": {},
+  "scripts": {
+    "test": "jest"
+  }
+}
